+++ src/main/java/calculator/application/ExtractService.java
@@ -1,9 +1,10 @@
 package calculator.application;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ExtractService {
 
-    List<Long> extractNumbers(List<String> data);
+    List<BigInteger> extractNumbers(List<String> data);
 
 }
